fix(e2e): move the credential health check to its own spec file
Playwright rejects test.use({ trace }) inside a describe group, because the option
forces a new worker, so the previous commit failed every e2e leg at collection
time. File scope is the narrowest scope the option allows.

Keeping the test in its own file leaves trace capture on for the rest of
configEditor.spec.ts. The full test title is unchanged.
