+++ tests/e2e/configEditor.spec.ts
@@ -1,5 +1,4 @@
 import { expect, test } from '@grafana/plugin-e2e';
-import { Page } from '@playwright/test';
 
 import { type CloudWatchJsonData } from '../../src/types';
 
@@ -9,14 +8,6 @@ const PROVISIONED_FILE = 'datasources.yml';
 // GRAFANA_URL is set only by the Cloud cron workflow (playwright-cloud); its presence signals
 // a run against the shared Cloud instance rather than local/PR CI.
 const isCloudRun = !!process.env.GRAFANA_URL;
-
-// Selects a Private Data Source Connect network in the datasource config editor. The
-// combobox is a Grafana-core element, present only when PDC is available on the instance
-// (i.e. the shared Cloud instance), so it is called only when a network name is injected.
-async function configurePDC(page: Page, networkName: string) {
-  await page.getByRole('combobox', { name: 'Private data source connect' }).click();
-  await page.getByText(networkName).click();
-}
 
 test.describe('Config editor', () => {
   test.describe('rendering', () => {
@@ -95,44 +86,7 @@ test.describe('Config editor', () => {
       await expect(configPage).toHaveAlert('error');
     });
 
-    // Trace capture is off here because a trace records the request body that carries the
-    // injected credentials. They also stay out of the DOM, because the datasource is created
-    // through the API: core renders Secret Access Key as a plain input, not a password input.
-    test.describe('injected credentials', () => {
-      test.use({ trace: 'off' });
-
-      test(
-        'valid injected credentials should pass the health check',
-        { tag: '@aws' },
-        async ({ createDataSource, gotoDataSourceConfigPage, page }) => {
-          // Consumes the AWS test-account credentials injected by the Cloud cron workflow
-          // (playwright-cloud) from the data-sources Vault mount as AWS_ACCESS_KEY_ID /
-          // AWS_SECRET_ACCESS_KEY / AWS_DEFAULT_REGION. Skipped when they are absent (fork PRs,
-          // or local runs with no Vault access).
-          const accessKey = process.env.AWS_ACCESS_KEY_ID;
-          const secretKey = process.env.AWS_SECRET_ACCESS_KEY;
-          const region = process.env.AWS_DEFAULT_REGION;
-          test.skip(
-            !accessKey || !secretKey || !region,
-            'Requires the injected AWS credentials (Cloud cron / CI Vault)'
-          );
-
-          // Mirrors provisioning/datasources/datasources.yml, which is the same auth shape.
-          const ds = await createDataSource({
-            type: PLUGIN_TYPE,
-            jsonData: { authType: 'keys', defaultRegion: region ?? '' },
-            secureJsonData: { accessKey: accessKey ?? '', secretKey: secretKey ?? '' },
-          });
-          const configPage = await gotoDataSourceConfigPage(ds.uid);
-
-          if (process.env.DS_PDC_NETWORK_NAME) {
-            await configurePDC(page, process.env.DS_PDC_NETWORK_NAME);
-          }
-
-          const response = await configPage.saveAndTest();
-          expect(response.ok()).toBe(true);
-        }
-      );
-    });
+    // The health check with real injected credentials lives in configEditorCredentials.spec.ts,
+    // which turns trace capture off at file scope.
   });
 });
